@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:sas/app/shared/myTheme.dart';
 
+
 import '../sizeConfig.dart';
 
 class BText extends StatelessWidget {
@@ -11,18 +12,18 @@ class BText extends StatelessWidget {
   double size;
   TextAlign align;
 
-  BText(this.texto, {this.cor, this.fweight, this.size, this.align});
+  BText(this.texto, {this.cor = MyTheme.greyText, this.fweight = FontWeight.normal, this.size = 14, this.align = TextAlign.center});
   @override
   Widget build(BuildContext context) {
     var sc = SizeConfig.of(context);
     return Text(
       texto,
-      textAlign: align == null ? TextAlign.center : align,
+      textAlign: align,
       style: GoogleFonts.poppins(
         textStyle: TextStyle(
-          color: cor == null ? MyTheme().getGreyTextColor() : cor,
-          fontWeight: fweight == null ? FontWeight.normal : fweight,
-          fontSize: size == null ? sc.getSize(14) : sc.getSize(size),
+          color: cor,
+          fontWeight: fweight,
+          fontSize: sc.getSize(size),
         ),
       ),
     );
